int32 -> uint32

+++ src/IndexableFile.h
@@ -128,7 +128,7 @@ public:
                         break;
 
                     case loop_type::looptype_repeat:
-                        if (nextIdx > _loop_finish_blocks)  // trying to go too far...
+                        if (nextIdx > (int)_loop_finish_blocks)  // trying to go too far...
                         {
                             /*
                             indexedbuffer* nl = _buffers.back();
@@ -143,7 +143,7 @@ public:
                         break;
 
                     case loop_type::looptype_pingpong: // forwards, we're pinging
-                        if (nextIdx > _loop_finish_blocks)  // trying to go too far...
+                        if (nextIdx > (int)_loop_finish_blocks)  // trying to go too far...
                         {
                             bufferAction = nopBuffer;
                             if (read == _buffers[MAX_NUM_BUFFERS - 1]->status)  // last update used last buffer, so...
@@ -167,7 +167,7 @@ public:
                         break;
 
                     case loop_type::looptype_repeat:
-                        if (nextIdx < _loop_start_blocks)   // trying to go too far...
+                        if (nextIdx < (int)_loop_start_blocks)   // trying to go too far...
                         {
                             int nextLast = _buffers[MAX_NUM_BUFFERS - 1]->index;
                             if (nextLast - 1 == nextIdx)    // haven't looped back already...
@@ -178,7 +178,7 @@ public:
                         break;
 
                     case loop_type::looptype_pingpong: // reverse, we're ponging
-                        if (nextIdx < _loop_start_blocks)   // trying to go too far...
+                        if (nextIdx < (int)_loop_start_blocks)   // trying to go too far...
                         {
                             bufferAction = nopBuffer;
                             if (read == _buffers[MAX_NUM_BUFFERS - 1]->status)  // last update used last buffer, so...
@@ -444,8 +444,8 @@ public:
     }
 
     void setLoopType(loop_type l) { _loop_type = l; }
-    void setLoopStart(int32_t l) { _loop_start_blocks = l >> buffer_to_index_shift; }
-    void setLoopFinish(int32_t l) { _loop_finish_blocks = l >> buffer_to_index_shift; }
+    void setLoopStart(uint32_t l) { _loop_start_blocks = l >> buffer_to_index_shift; }
+    void setLoopFinish(uint32_t l) { _loop_finish_blocks = l >> buffer_to_index_shift; }
 
 protected:
     TFile _file;
@@ -454,8 +454,8 @@ protected:
     // we need a copy of the loop parameters in
     // order to pre-load the buffers correctly
     loop_type _loop_type = loop_type::looptype_none;
-    int32_t _loop_start_blocks = 0;
-    int32_t _loop_finish_blocks = 0;
+    uint32_t _loop_start_blocks = 0;
+    uint32_t _loop_finish_blocks = 0;
 
     bool _bufInPSRAM = false;
